Use data as the key.
This makes it a lot easier to remove a point afer adding it: you can use the
same data you used to add the point originally, rather than needing to keep a
reference to the point returned by quadtree.add. Now, quadtree.add and
quadtree.remove return the quadtree to facilitate method chaining.